Refactor to use JS build tools
