+++ src/index.js
@@ -1,0 +1,5 @@
+import scripts from './scripts';
+
+export default {
+  scripts,
+};
