+++ src/scripts/index.js
@@ -1,0 +1,9 @@
+import { patient } from './patient';
+import { lab_test } from './lab_test';
+import { sales_invoice } from './sales_invoice';
+
+export default {
+  patient,
+  lab_test,
+  sales_invoice,
+};
